Revert "workspace activation service"
This reverts commit 782f28b67fa34fa3735daf62ccb8acbb55d26861.
